Created prototype login screen, improved registration input validation, set up program infrastructure further
Functionality to insert user details in login screen to access another part of the UI is now in the code. Iterates through the arrays holding user objects to see if the inputted details match any object stored in the array.

Extra input validation inserted for the registration form. '@' now cannot be used as the first character of a username, as this will be reserved for admin accounts. 'Back' button functionality also added, so user can now navigate back to login screen.

Set up additional classes such as 'VehicleManager' (similar in function to UserManager) and loginSystem (so I can attempt to implement a design pattern into the login system, to tell the system what UI to display based on if it is a customer or admin logging in).