Login system completed
System can now log into different parts of the UI, depending on user input determining if they're an admin or not (ie. an @ at the start of their username). Also added a default admin account that is created upon opening the app; account details to be added to the readme.